Move recent contract methods into it's own repo
