+++ src/Infrastructure/Databases/EFUserRepository.cs
@@ -11,7 +11,7 @@ namespace Infrastructure.Databases;
 /// <summary>
 /// Stores and fetches users from an Entity Framework Core database.
 /// </summary>
-public sealed class EFUserRepository : IUserRepository, IRecentContractRepository
+public sealed class EFUserRepository : IUserRepository
 {
     private const string AdminUserName = "admin";
 
@@ -40,46 +40,6 @@ public sealed class EFUserRepository : IUserRepository, IRecentContractRepositor
         _ = Users.Add(user);
         _ = _context.SaveChanges();
         _logger.LogInformation("Added a new user with name {Name} and id {Id} to the database", user.Name, user.Id);
-    }
-
-    /// <inheritdoc />
-    public void AddRecent(string username, Contract contract)
-    {
-        User user = GetUserByUserName(username);
-
-        RecentlyViewedContract? recentlyViewedContract =
-            user.RecentlyViewContracts.SingleOrDefault(recentContract => recentContract.ContractId == contract.Id);
-
-        // Check whether to create a new recent contract or just update the viewed time.
-        if (recentlyViewedContract == null)
-        {
-            var recentContract = new RecentlyViewedContract(contract.Id, user.Id);
-            user.RecentlyViewContracts.Add(recentContract);
-        }
-        else
-        {
-            recentlyViewedContract.LastViewed = DateTime.Now;
-        }
-
-        try
-        {
-            _context.SaveChanges();
-        }
-        catch (DataException e)
-        {
-            _logger.LogError("Could not add contract to recently viewed for user to database: {Message}", e.Message);
-        }
-    }
-
-    /// <inheritdoc />
-    public void RemoveRecent(RecentlyViewedContract contractToRemove)
-    {
-        _context.RecentlyViewedContracts.Remove(contractToRemove);
-        _ = _context.SaveChanges();
-        _logger.LogInformation(
-            "Removed contract with id {ContractId} from user with id {UserId}",
-            contractToRemove.ContractId,
-            contractToRemove.UserId);
     }
 
     /// <inheritdoc />
@@ -112,13 +72,6 @@ public sealed class EFUserRepository : IUserRepository, IRecentContractRepositor
     }
 
     /// <inheritdoc />
-    public IList<RecentlyViewedContract> FetchRecentContracts(string id)
-    {
-        User user = GetUserByUserName(id);
-        return user.RecentlyViewContracts;
-    }
-
-    /// <inheritdoc />
     public User? Fetch(string username)
     {
         return Users.FirstOrDefault(user => user.Name == username);
@@ -143,11 +96,5 @@ public sealed class EFUserRepository : IUserRepository, IRecentContractRepositor
         {
             _logger.LogError("Could not create admin: {Message}", e.Message);
         }
-    }
-
-    private User GetUserByUserName(string username)
-    {
-        User user = Users.Include(user => user.RecentlyViewContracts).First(user => user.Name == username);
-        return user;
     }
 }
