+++ src/Infrastructure/InjectionExtensions.cs
@@ -40,7 +40,7 @@ public static class InjectionExtensions
                    ServiceLifetime.Transient)
                .AddScoped<IContractRepository, EFContractRepository>()
                .AddScoped<IUserRepository, EFUserRepository>()
-               .AddScoped<IRecentContractRepository, EFUserRepository>()
+               .AddScoped<IRecentContractRepository, EFRecentContractRepository>()
                .AddSingleton<IStatusUpdateRepository, InMemoryStatusUpdateRepository>()
                .AddSingleton<IImageRepository, LocalFileRepository>(provider =>
                {
