+++ tests/Infrastructure.Tests/Infrastructure.Tests.csproj
@@ -48,8 +48,4 @@
     <ProjectReference Include="..\..\src\Infrastructure\Infrastructure.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Contracts" />
-  </ItemGroup>
-
 </Project>
